done for now

+++ src/main/java/com/ideas2it/healthcare/response/SuccessResponse.java
@@ -67,7 +67,7 @@ public class SuccessResponse {
      * @return {@link ResponseEntity<Map<String, Object>>}
      */
     public ResponseEntity<Map<String, Object>> responseEntity
-            (String message, List<?> entityList, HttpStatus status, double totalPages) {
+            (String message, List<?> entityList, HttpStatus status, Integer totalPages) {
         Map<String, Object> response = new LinkedHashMap<>();
         response.put(Constants.MESSAGE, message);
         response.put(Constants.ENTITY, null);
