+++ src/main/java/com/ideas2it/healthcare/util/MathUtil.java
@@ -2,8 +2,8 @@ package com.ideas2it.healthcare.util;
 
 public class MathUtil {
 
-    public static Integer getExactCount(Integer count, Integer rows) {
-        float result = (float) (count/rows);
+    public static Integer getExactCount(float count, float rows) {
+        float result =  count/rows;
         if (Math.floor(result) < result) {
             return Math.round(result) + 1;
         }
